requirements.txt
Django>=3.0,<3.1
psycopg2>=2.8,<2.9
django-recaptcha>=2.0.6,<2.1
pillow>=7.0<7.1
django-taggit>=1.2<1.3
requests>=2.22<2.23
backports-datetime-fromisoformat>=1.0<1.1
django-filebrowser-no-grappelli>=3.8<3.9
django-streamfield>=1.2<1.3
django-crispy-forms>=1.8<1.9
